Move IQueryStageAggregator to query package
