+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/IQueryStageAggregatorVisitor.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.api;
+package org.bithon.server.storage.datasource.query;
 
 /**
  * @author Frank Chen
