+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/QueryStageAggregators.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.api;
+package org.bithon.server.storage.datasource.query;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
