+++ server/storage/src/main/java/org/bithon/server/storage/datasource/spec/sum/SumMetricSpec.java
@@ -19,8 +19,8 @@ package org.bithon.server.storage.datasource.spec.sum;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import org.bithon.server.storage.datasource.DataSourceSchema;
-import org.bithon.server.storage.datasource.api.IQueryStageAggregator;
-import org.bithon.server.storage.datasource.api.QueryStageAggregators;
+import org.bithon.server.storage.datasource.query.IQueryStageAggregator;
+import org.bithon.server.storage.datasource.query.QueryStageAggregators;
 import org.bithon.server.storage.datasource.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.spec.IMetricSpecVisitor;
